+++ SLIDE-x-BENCH/KERNEL/gcd/frst.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include <values.h>
 
-typedef double TARGET_TYPE;
+typedef int64_t TARGET_TYPE;
 typedef uint64_t TARGET_INDEX;
 
 TARGET_TYPE modulo(TARGET_TYPE x, TARGET_TYPE y)
